third_party/WebKit/LayoutTests/fast/css/usecounter-negative-background-size-longhand.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
    let NegativeBackgroundSize = 2402; // From web_feature.mojom

    let isCounted = () => internals.isUseCounted(document, NegativeBackgroundSize);
    var div = document.createElement('div');

    div.style = 'background-size: 1px 2px;';
    div.style = '-webkit-mask-size: -1px -2px;';
    assert_false(isCounted(),
                 '-webkit-mask-size should not be counted');

    div.style = 'background-size: 1px -2px;';
    assert_true(isCounted(),
                'background-size should be counted');
}, 'Negative size is use counted for background-size');
</script>
